shuffle settings around
