+++ aiuser/settings/prompt.py
@@ -5,6 +5,7 @@ from typing import Optional, Union
 import discord
 import tiktoken
 from redbot.core import checks, commands
+from redbot.core.utils.chat_formatting import box, pagify
 from redbot.core.utils.menus import SimpleMenu, start_adding_reactions
 from redbot.core.utils.predicates import ReactionPredicate
 
@@ -214,40 +215,65 @@ class PromptSettings(MixinMeta):
         embed.add_field(name="Tokens", value=await self.get_tokens(ctx, prompt))
         return await ctx.send(embed=embed)
 
-    @prompt.group()
-    @checks.is_owner()
-    async def history(self, _):
-        """ Change the prompt context settings for the current server
-
-            The most recent messages that are within the time gap and message limits are used to create context.
-            Context is used to help the LLM generate a response.
-        """
+    @prompt.group(name="topics")
+    async def random_topics(self, _):
+        """ Manage topics to be used in random messages for current server (if enabled for server by bot owner) """
         pass
 
-    @history.command(name="backread", aliases=["messages", "size"])
-    @checks.is_owner()
-    async def history_backread(self, ctx: commands.Context, new_value: int):
-        """ Set max amount of messages to be used """
-        await self.config.guild(ctx.guild).messages_backread.set(new_value)
+    @random_topics.command(name="show", aliases=["list"])
+    async def show_random_topics(self, ctx: commands.Context):
+        """ Lists topics to used in random messages """
+        topics = await self.config.guild(ctx.guild).random_messages_topics()
+
+        if not topics:
+            return await ctx.send("The topic list is empty.")
+
+        formatted_list = "\n".join(f"{index+1}. {topic}" for index, topic in enumerate(topics))
+        pages = []
+        for text in pagify(formatted_list, page_length=888):
+            page = discord.Embed(
+                title=f"List of random message topics in {ctx.guild.name}",
+                description=box(text),
+                color=await ctx.embed_color())
+            pages.append(page)
+
+        if len(pages) == 1:
+            return await ctx.send(embed=pages[0])
+
+        for i, page in enumerate(pages):
+            page.set_footer(text=f"Page {i+1} of {len(pages)}")
+
+        return await SimpleMenu(pages).start(ctx)
+
+    @random_topics.command(name="add", aliases=["a"])
+    async def add_random_topics(self, ctx: commands.Context, *, topic: str):
+        """ Add a new topic """
+        topics = await self.config.guild(ctx.guild).random_messages_topics()
+        if topic in topics:
+            return await ctx.send("That topic is already in the list.")
+        if topic and len(topic) > await self.config.max_topic_length() and not await ctx.bot.is_owner(ctx.author):
+            return await ctx.send(f"Topic too long. Max length is {await self.config.max_topic_length()} characters.")
+        topics.append(topic)
+        await self.config.guild(ctx.guild).random_messages_topics.set(topics)
         embed = discord.Embed(
-            title="The number of previous messages used for context on this server is now:",
-            description=f"{new_value}",
+            title="Added topic to random message topics:",
+            description=f"{topic}",
             color=await ctx.embed_color())
+        embed.add_field(name="Tokens", value=await self.get_tokens(ctx, topic))
         return await ctx.send(embed=embed)
 
-    @history.command(name="time", aliases=["gap"])
-    @checks.is_owner()
-    async def history_time(self, ctx: commands.Context, new_value: int):
-        """ Set max time (s) allowed between messages to be used
-
-            eg. if set to 60, once messsages are more than 60 seconds apart, more messages will not be added.
-
-            Helpful to prevent the LLM from mixing up context from different conversations.
-        """
-        await self.config.guild(ctx.guild).messages_backread_seconds.set(new_value)
+    @random_topics.command(name="remove", aliases=["rm", "delete"])
+    async def remove_random_topics(self, ctx: commands.Context, *, number: int):
+        """ Removes a topic (by number) from the list"""
+        topics = await self.config.guild(ctx.guild).random_messages_topics()
+        if not (1 <= number <= len(topics)):
+            return await ctx.send("Invalid topic number.")
+        topic = topics[number - 1]
+        topics.remove(topic)
+        await self.config.guild(ctx.guild).random_messages_topics.set(topics)
         embed = discord.Embed(
-            title="The max time (s) allowed between messages for context on this server is now:",
-            description=f"{new_value}",
+            title="Removed topic from random message topics:",
+            description=f"{topic}",
             color=await ctx.embed_color())
         return await ctx.send(embed=embed)
