+++ aiuser/settings/triggers.py
@@ -100,64 +100,3 @@ class TriggerSettings(MixinMeta):
             description=f"{percent:.2f}%",
             color=await ctx.embed_color())
         return await ctx.send(embed=embed)
-
-    @random.group(name="topics")
-    async def random_topics(self, _):
-        """ Manage topics to be used in random messages for current server """
-        pass
-
-    @random_topics.command(name="show", aliases=["list"])
-    async def show_random_topics(self, ctx: commands.Context):
-        """ Lists topics to used in random messages """
-        topics = await self.config.guild(ctx.guild).random_messages_topics()
-
-        if not topics:
-            return await ctx.send("The topic list is empty.")
-
-        formatted_list = "\n".join(f"{index+1}. {topic}" for index, topic in enumerate(topics))
-        pages = []
-        for text in pagify(formatted_list, page_length=888):
-            page = discord.Embed(
-                title=f"List of random message topics in {ctx.guild.name}",
-                description=box(text),
-                color=await ctx.embed_color())
-            pages.append(page)
-
-        if len(pages) == 1:
-            return await ctx.send(embed=pages[0])
-
-        for i, page in enumerate(pages):
-            page.set_footer(text=f"Page {i+1} of {len(pages)}")
-
-        return await SimpleMenu(pages).start(ctx)
-
-    @random_topics.command(name="add", aliases=["a"])
-    async def add_random_topics(self, ctx: commands.Context, *, topic: str):
-        """ Add a new topic """
-        topics = await self.config.guild(ctx.guild).random_messages_topics()
-        if topic in topics:
-            return await ctx.send("That topic is already in the list.")
-        if topic and len(topic) > await self.config.max_topic_length() and not await ctx.bot.is_owner(ctx.author):
-            return await ctx.send(f"Topic too long. Max length is {await self.config.max_topic_length()} characters.")
-        topics.append(topic)
-        await self.config.guild(ctx.guild).random_messages_topics.set(topics)
-        embed = discord.Embed(
-            title="Added topic to random message topics:",
-            description=f"{topic}",
-            color=await ctx.embed_color())
-        return await ctx.send(embed=embed)
-
-    @random_topics.command(name="remove", aliases=["rm", "delete"])
-    async def remove_random_topics(self, ctx: commands.Context, *, number: int):
-        """ Removes a topic (by number) from the list"""
-        topics = await self.config.guild(ctx.guild).random_messages_topics()
-        if not (1 <= number <= len(topics)):
-            return await ctx.send("Invalid topic number.")
-        topic = topics[number - 1]
-        topics.remove(topic)
-        await self.config.guild(ctx.guild).random_messages_topics.set(topics)
-        embed = discord.Embed(
-            title="Removed topic from random message topics:",
-            description=f"{topic}",
-            color=await ctx.embed_color())
-        return await ctx.send(embed=embed)
